pyproject.toml
[tool.poetry]
name = "StallmansDiscord"
version = "0.1.0"
description = "Discord version of Not thing, GNU/thing"
authors = ["Ian P <ian@aotc.us>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.1"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.1"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
